requirements.txt
#apex==0.1
backcall==0.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.6.1
colorama==0.4.1
cycler==0.10.0
Cython==0.29.7
cytoolz==0.9.0.1
dask==1.0.0
decorator==4.3.0
docopt==0.6.2
easydict==1.9
h5py==2.9.0
idna==2.8
imageio==2.4.1
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
jsonpickle==0.9.6
kiwisolver==1.0.1
llvmlite==0.27.0
#-e git+https://github.com/facebookresearch/maskrcnn-benchmark.git@55796a04ea770029a80cf5933cc5c3f3f6fa59cf#egg=maskrcnn_benchmark
matplotlib==3.0.2
#mkl-fft==1.0.10
#mkl-random==1.0.2
munch==2.3.2
networkx==2.2
ninja==1.9.0
numba==0.42.0
numpy==1.15.4
olefile==0.46
opencv-python==3.4.4.19
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py-cpuinfo==4.0.0
#pycocotools==2.0
pycparser==2.19
Pygments==2.6.1
pymongo==3.7.2
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
requests==2.21.0
sacred==0.7.4
scikit-image==0.14.1
scipy==1.2.0
six==1.12.0
toolz==0.9.0
torch==1.0.0
torchfile==0.1.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.29.1
traitlets==4.3.3
typing==3.6.4
urllib3==1.24.1
visdom==0.1.8.8
wcwidth==0.1.9
websocket-client==0.54.0
wrapt==1.11.1
yacs==0.1.6
